Add “add member” primary admin action link to the group discussion listing view that appears on group detail page.
This action link existed in the group module (in
group.links.action.yml) and referenced the group_members View. Since we
override the group_members View with discussions_group_members we had
to create the action link in discussions.links.action.yml and have it
appear on our discussions_group_members View page.

Resolves #72