Make email generated from comments in the same discussion maintain threading.
To do this we leverage field_discussions_message_id, which exists on
both the public and private discussion entities, to store the message
ID of the last comment on the discussion thread. Note: the value in the
message ID field matches the Message-Id email header.

We then use the Message-Id to set the In-Reply-To email header for the
next message sent out by the system.